[GB-6] rename: rename domain to entity
