+++ src/main/java/garlicbears/_quiz/domain/user/repository/UserRepository.java
@@ -1,6 +1,6 @@
 package garlicbears._quiz.domain.user.repository;
 
-import garlicbears._quiz.domain.user.domain.User;
+import garlicbears._quiz.domain.user.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
