+++ src/main/java/garlicbears/_quiz/global/config/auth/PrincipalDetailsService.java
@@ -1,6 +1,6 @@
 package garlicbears._quiz.global.config.auth;
 
-import garlicbears._quiz.domain.user.domain.User;
+import garlicbears._quiz.domain.user.entity.User;
 import garlicbears._quiz.domain.user.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.core.userdetails.UserDetails;
