+++ src/main/java/garlicbears/_quiz/global/entity/Active.java
@@ -1,4 +1,4 @@
-package garlicbears._quiz.global.domain;
+package garlicbears._quiz.global.entity;
 
 public enum Active {
     active,
